tests/requirements-test.txt
# tests/requirements-test.txt
pytest==7.4.3
pytest-html==4.1.1
requests==2.31.0
psycopg2-binary==2.9.9
six==1.17.0
pydantic==2.5.0

aiokafka==0.10.0
kafka-python-ng